Changed the hardsid availability checking routine to return an int instead of an unsigned int.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@25862 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/win32/hardsid.c
@@ -242,13 +242,13 @@ void hardsid_set_machine_parameter(long cycles_per_sec)
 {
 }
 
-unsigned int hardsid_available(void)
+int hardsid_available(void)
 {
     if (init_interface() < 0) {
         return 0;
     }
 
-    return (unsigned int)GetHardSIDCount();
+    return GetHardSIDCount();
 }
 
 void hardsid_set_device(unsigned int chipno, unsigned int device)
